Support PHP user functions with references
PHP7 supports calls to named functions (this or static) with
call_user_func. However, call_user_func does not support reference
variables as arguments. The workaround is to use call_user_func_array
instead of call_user_func.

Added the phpPHP:delegateReferenceSupport markup on a delegateCall or
a static call with a plxPHP:callExpression to force the generator to
emit call_user_array_func instead of call_user_array. Also added
plxPHP:reference support to a plx:passParam to support passing
reference arguments to the array generated for this structure,
where the PHP behavior is to pass a value instead of a reference,
even if the passed variable is a reference type.