Add VS2019 build support
Note that the VS2017-built packages also installs in VS2019. VS2005
through VS2015 are built in one msi-based package, and we should be
able to build VS2017 and higher with a single .vsix. This is meant
as convenience if we need to build with VS2019.

Cleaned up an error in VS2005 build.