Enable PLiX generation with no byte order mark
Some languages (specifically PHP) do not support byte order marks, but
this is the default output for XSL-generated output. Enable
configuration to specify that output for specific language transforms
should not generate the byte order mark on the file.

For versions through VS2015 this is done in the PlixSettings.xml file
with the noByteOrderMark attribute on the formatter element. For
VS2017 and VS2019 the PLiX/FormatterOptions key is added in parallel
to the PLiX/Formatters key. If Formatters@php is set to a transform
file then FormatterOptions/php@NoByteOrderMark can be set to 1 to turn
off the generation of a byte order mark.