Git versioning and per-user install
Make versioning on all files git-based.

With VSIXPerUser=1 environment variable set, create a per-user
installation. This does not install schema files with VS, but it
does adjust the schema catalog file so that this can be opened
directly in Visual Studio, which causes the XML editor to recognize
the schema files listed in the catalog.